leetcode, refactoring
